refactor(auth): auth 관련 configuration 패키지 auth 아래로 이동
