+++ backend/src/main/java/codezap/auth/configuration/AuthArgumentResolver.java
@@ -1,4 +1,4 @@
-package codezap.member.configuration;
+package codezap.auth.configuration;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -8,13 +8,16 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-import codezap.member.service.AuthService;
+import codezap.auth.manager.CredentialManager;
+import codezap.auth.provider.CredentialProvider;
+import codezap.member.dto.MemberDto;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
 
-    private final AuthService authService;
+    private final CredentialManager credentialManager;
+    private final CredentialProvider credentialProvider;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
@@ -22,13 +25,14 @@ public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
     }
 
     @Override
-    public Object resolveArgument(
+    public MemberDto resolveArgument(
             MethodParameter parameter,
             ModelAndViewContainer mavContainer,
             NativeWebRequest webRequest,
             WebDataBinderFactory binderFactory
     ) {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-        return authService.authorizeByCookie(request.getCookies());
+        String credential = credentialManager.getCredential(request);
+        return MemberDto.from(credentialProvider.extractMember(credential));
     }
 }
