+++ backend/src/main/java/codezap/auth/configuration/AuthWebConfiguration.java
@@ -1,4 +1,4 @@
-package codezap.member.configuration;
+package codezap.auth.configuration;
 
 import java.util.List;
 
@@ -6,17 +6,19 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import codezap.member.service.AuthService;
+import codezap.auth.manager.CredentialManager;
+import codezap.auth.provider.CredentialProvider;
 import lombok.RequiredArgsConstructor;
 
 @Configuration
 @RequiredArgsConstructor
 public class AuthWebConfiguration implements WebMvcConfigurer {
 
-    private final AuthService authService;
+    private final CredentialProvider credentialProvider;
+    private final CredentialManager credentialManager;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new AuthArgumentResolver(authService));
+        resolvers.add(new AuthArgumentResolver(credentialManager, credentialProvider));
     }
 }
